server/requirements/base.txt
fastapi
pydantic
pydantic_settings
pydantic[email]
uvicorn
passlib[bcrypt]
python-jose
cryptography
pycryptodome
python-multipart